refactor/#61: global.aws -> infra.lambda 패키지로 리팩토링
